Move /check modules into their own folder
