+++ app/controllers/return-requirements-setup.controller.js
@@ -10,7 +10,7 @@ const AddService = require('../services/return-requirements/setup/add.service.js
 const AdditionalSubmissionOptionsService = require('../services/return-requirements/setup/additional-submission-options.service.js')
 const AgreementsExceptionsService = require('../services/return-requirements/setup/agreements-exceptions.service.js')
 const CancelService = require('../services/return-requirements/setup/cancel.service.js')
-const CheckService = require('../services/return-requirements/setup/check.service.js')
+const CheckService = require('../services/return-requirements/setup/check/check.service.js')
 const DeleteNoteService = require('../services/return-requirements/setup/delete-note.service.js')
 const ExistingService = require('../services/return-requirements/setup/existing.service.js')
 const FeatureFlagsConfig = require('../../config/feature-flags.config.js')
@@ -30,7 +30,7 @@ const SubmitAbstractionPeriod = require('../services/return-requirements/setup/s
 const SubmitAdditionalSubmissionOptionsService = require('../services/return-requirements/setup/submit-additional-submission-options.service.js')
 const SubmitAgreementsExceptions = require('../services/return-requirements/setup/submit-agreements-exceptions.service.js')
 const SubmitCancel = require('../services/return-requirements/setup/submit-cancel.service.js')
-const SubmitCheckService = require('../services/return-requirements/setup/submit-check.service.js')
+const SubmitCheckService = require('../services/return-requirements/setup/check/submit-check.service.js')
 const SubmitExistingService = require('../services/return-requirements/setup/submit-existing.service.js')
 const SubmitFrequencyCollectedService = require('../services/return-requirements/setup/submit-frequency-collected.service.js')
 const SubmitFrequencyReportedService = require('../services/return-requirements/setup/submit-frequency-reported.service.js')
