+++ test/controllers/return-requirements-setup.controller.test.js
@@ -16,7 +16,7 @@ const AbstractionPeriodService = require('../../app/services/return-requirements
 const AdditionalSubmissionOptionsService = require('../../app/services/return-requirements/setup/additional-submission-options.service.js')
 const AgreementsExceptionService = require('../../app/services/return-requirements/setup/agreements-exceptions.service.js')
 const CancelService = require('../../app/services/return-requirements/setup/cancel.service.js')
-const CheckService = require('../../app/services/return-requirements/setup/check.service.js')
+const CheckService = require('../../app/services/return-requirements/setup/check/check.service.js')
 const DeleteNoteService = require('../../app/services/return-requirements/setup/delete-note.service.js')
 const ExistingService = require('../../app/services/return-requirements/setup/existing.service.js')
 const FrequencyCollectedService = require('../../app/services/return-requirements/setup/frequency-collected.service.js')
